Refactor: 댓글 가져오는 api, comment 디렉토리로 이전
- article comment에 댓글 가져오는 controller가 있어서 의존성이 이상하게 흐르는 문제 수정
- get-comment api를 위한 별도의 컨트롤러 추가
